o/restart: restructure a bit and move immediate
Move the "immediate" handling back to where it was before, restructure the new restart code a bit and seperate out the new class. Rename to RestartContext.
